Modified math to match how Web-UI handles presets 0-7 rather then 1-8
The web-ui of the camera uses a zero starting point for its presets.
Changed the math to subtract one from the preset being sent via zoneminder.
This allows both ZM and Web-UI to have matching presets, this should also
have the affect of having ZoneMinder goto the correct Preset_1 if setup through WebUI.

This will throw off any preset's currently in ZM, could use the web-ui of camrea to move
2->1
3->2
4->3
5->4
6->5
7->6
Will need to move 8->7 within ZM before using this new script, as this is stored in a 9 slot